// src/shared/api/autoGenerateTypes/orderDtoStatus.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * 스터디 메이트 API 문서 (Dev)
 * 스터디메이트 BFF Server API 문서 (Dev)
 * OpenAPI spec version: v1
 */

export type OrderDtoStatus =
  (typeof OrderDtoStatus)[keyof typeof OrderDtoStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OrderDtoStatus = {
  REQUEST: 'REQUEST',
  PAID: 'PAID',
  PENDING: 'PENDING',
  REQ_CANCELLED: 'REQ_CANCELLED',
  PAY_ALL_CANCELLED: 'PAY_ALL_CANCELLED',
  PAY_PARTIAL_CANCELLED: 'PAY_PARTIAL_CANCELLED',
  FAILED: 'FAILED',
} as const;